+++ models/Configuration.php
@@ -10,8 +10,10 @@ namespace humhub\modules\cleanTheme\models;
 
 use humhub\components\SettingsManager;
 use humhub\modules\cleanTheme\helpers\ColorHelper;
+use humhub\widgets\Button;
 use Yii;
 use yii\base\Model;
+use yii\helpers\Html;
 use yii\helpers\Inflector;
 
 class Configuration extends Model
@@ -21,7 +23,7 @@ class Configuration extends Model
     public const SUPPORTED_LESS_FUNCTIONS = ['darken', 'lighten', 'fade', 'fadein', 'fadeout'];
     public const UNSUPPORTED_LESS_FUNCTIONS = ['saturate', 'desaturate', 'spin'];
     public const DYNAMIC_CSS_FILE_PATH = '@clean-theme/resources/css/';
-    public const DYNAMIC_CSS_FILE_NAME = 'humhub.clean-theme.css-variables.css';
+    public const DYNAMIC_CSS_FILE_NAME = 'humhub.clean-theme.dynamic.css';
     public const CSS_ATTRIBUTE_UNITS = [
         'containerMaxWidth' => 'px',
         'topMenuNavJustifyContent' => '',
@@ -46,6 +48,7 @@ class Configuration extends Model
         'backgroundColorSecondary' => '',
         'backgroundColorPage' => '',
         'backgroundColorHighlight' => '',
+        'backgroundColorHighlightSoft' => '',
         'fontFamily' => '',
         'headingFontFamily' => '',
         'fontSize' => 'px',
@@ -70,7 +73,7 @@ class Configuration extends Model
 
     public SettingsManager $settingsManager;
 
-    public int $containerMaxWidth = 1600;
+    public string $containerMaxWidth = '1600';
     public string $topMenuNavJustifyContent = 'center';
     public string $default = '#f3f3f3';
     public string $primary = '#31414a';
@@ -92,27 +95,28 @@ class Configuration extends Model
     public string $backgroundColorMain = '#ffffff';
     public string $backgroundColorSecondary = '#f7f7f7';
     public string $backgroundColorPage = '#f1f4f5';
-    public string $backgroundColorHighlight = '#fff8e0';
-    public ?string $fontFamily = null;
-    public ?string $headingFontFamily = null;
-    public int $fontSize = 13;
-    public int $phFontSize = 16; // Panel Heading
-    public int $rtFontSize = 13; // Reach Text
-    public float $h1FontSize = 1.7;
-    public float $h1RtFontSize = 1.45;
-    public float $h2FontSize = 1.5;
-    public float $h2RtFontSize = 1.3;
-    public float $h3FontSize = 1.2;
-    public float $h4FontSize = 1.1;
-    public float $h5FontSize = 1.0;
-    public float $h6FontSize = 0.85;
-    public int $fontWeight = 400;
-    public int $phFontWeight = 700;
-    public int $panelBorderWidth = 1;
+    public string $backgroundColorHighlight = '#daf0f3';
+    public string $backgroundColorHighlightSoft = '#f2f9fb';
+    public string $fontFamily = 'Open Sans';
+    public string $headingFontFamily = 'Open Sans';
+    public string $fontSize = '13';
+    public string $phFontSize = '16'; // Panel Heading
+    public string $rtFontSize = '13'; // Reach Text
+    public string $h1FontSize = '1.7';
+    public string $h1RtFontSize = '1.45';
+    public string $h2FontSize = '1.5';
+    public string $h2RtFontSize = '1.3';
+    public string $h3FontSize = '1.2';
+    public string $h4FontSize = '1.1';
+    public string $h5FontSize = '1.0';
+    public string $h6FontSize = '0.85';
+    public string $fontWeight = '400';
+    public string $phFontWeight = '700';
+    public string $panelBorderWidth = '1';
     public string $panelBorderStyle = 'solid';
     public string $panelBorderColor = '#c7c9e7';
-    public int $panelBorderRadius = 4;
-    public ?string $panelBoxShadow = null;
+    public string $panelBorderRadius = '4';
+    public ?string $panelBoxShadow = 'none';
 
     public static function getJustifyContentOptions(): array
     {
@@ -135,6 +139,11 @@ class Configuration extends Model
         ];
     }
 
+    private static function isFontAttribute($attributeName): bool
+    {
+        return str_ends_with($attributeName, 'ontFamily');
+    }
+
     /**
      * @inheritdoc
      */
@@ -145,7 +154,7 @@ class Configuration extends Model
             [['h1FontSize', 'h1RtFontSize', 'h2FontSize', 'h2RtFontSize', 'h3FontSize', 'h4FontSize', 'h5FontSize', 'h6FontSize'], 'number'],
             [['topMenuNavJustifyContent', 'default', 'primary', 'info', 'success', 'warning', 'danger', 'link', 'menuBorderColor', 'textColorHeading',
                 'textColorMain', 'textColorDefault', 'textColorSecondary', 'textColorHighlight', 'textColorSoft', 'textColorSoft2', 'textColorSoft3',
-                'textColorContrast', 'backgroundColorMain', 'backgroundColorSecondary', 'backgroundColorPage', 'backgroundColorHighlight', 'fontFamily', 'headingFontFamily', 'panelBorderStyle', 'panelBorderColor', 'panelBoxShadow'], 'string'],
+                'textColorContrast', 'backgroundColorMain', 'backgroundColorSecondary', 'backgroundColorPage', 'backgroundColorHighlight', 'backgroundColorHighlightSoft', 'fontFamily', 'headingFontFamily', 'panelBorderStyle', 'panelBorderColor', 'panelBoxShadow'], 'string'],
         ];
     }
 
@@ -166,18 +175,19 @@ class Configuration extends Model
             'link' => Yii::t('CleanThemeModule.config', 'Link main color'),
             'menuBorderColor' => Yii::t('CleanThemeModule.config', 'Menu border color'),
             'textColorHeading' => Yii::t('CleanThemeModule.config', 'Heading text color'),
-            'textColorMain' => Yii::t('CleanThemeModule.config', 'Main text color'),
-            'textColorDefault' => Yii::t('CleanThemeModule.config', 'Default text color'),
-            'textColorSecondary' => Yii::t('CleanThemeModule.config', 'Secondary text color'),
-            'textColorHighlight' => Yii::t('CleanThemeModule.config', 'Highlight text color'),
-            'textColorSoft' => Yii::t('CleanThemeModule.config', 'Soft text color'),
-            'textColorSoft2' => Yii::t('CleanThemeModule.config', 'Soft text color 2'),
-            'textColorSoft3' => Yii::t('CleanThemeModule.config', 'Soft text color 3'),
-            'textColorContrast' => Yii::t('CleanThemeModule.config', 'Contrast text color'),
-            'backgroundColorMain' => Yii::t('CleanThemeModule.config', 'Main background color'),
-            'backgroundColorSecondary' => Yii::t('CleanThemeModule.config', 'Secondary background color'),
-            'backgroundColorPage' => Yii::t('CleanThemeModule.config', 'Page background color'),
-            'backgroundColorHighlight' => Yii::t('CleanThemeModule.config', 'Highlight background color'),
+            'textColorMain' => Yii::t('CleanThemeModule.config', 'Default body text color'),
+            'textColorDefault' => Yii::t('CleanThemeModule.config', 'Default text color for default icons and buttons etc.'),
+            'textColorSecondary' => Yii::t('CleanThemeModule.config', 'Secondary text color for some icons and buttons etc.'),
+            'textColorHighlight' => Yii::t('CleanThemeModule.config', 'Highlight text color for text like some active links hovered links etc.'),
+            'textColorSoft' => Yii::t('CleanThemeModule.config', 'Soft text color for side information like dates, placeholder, some dropdown headers'),
+            'textColorSoft2' => Yii::t('CleanThemeModule.config', 'Soft text color 2 for other side information like wall entry links (like/comment), help blocks in forms etc.'),
+            'textColorSoft3' => Yii::t('CleanThemeModule.config', 'Soft text color 3 for gridview summary and installer'),
+            'textColorContrast' => Yii::t('CleanThemeModule.config', 'Contrast text color for @primary, @info, @success, @warning, @danger buttons etc.'),
+            'backgroundColorMain' => Yii::t('CleanThemeModule.config', 'Main background color which should be in contrast with main, secondary and other text colors'),
+            'backgroundColorSecondary' => Yii::t('CleanThemeModule.config', 'Secondary background color used beside others for tabs'),
+            'backgroundColorPage' => Yii::t('CleanThemeModule.config', 'Page background color for other ui components as comment box etc.'),
+            'backgroundColorHighlight' => Yii::t('CleanThemeModule.config', 'Informative highlighting background color (e.g. Comment Permalink, Shared Item, Calendar CurDay, Wiki Active Page Nav, Mail Module Speech Bubble)'),
+            'backgroundColorHighlightSoft' => Yii::t('CleanThemeModule.config', 'Informative soft highlighting background color (e.g. Wiki Active Category)'),
             'fontFamily' => Yii::t('CleanThemeModule.config', 'Font family'),
             'headingFontFamily' => Yii::t('CleanThemeModule.config', 'Heading font family'),
             'fontSize' => Yii::t('CleanThemeModule.config', 'Font size'),
@@ -206,23 +216,35 @@ class Configuration extends Model
      */
     public function attributeHints()
     {
+        $inPx = Yii::t('CleanThemeModule.config', 'In pixel');
+        $inEm = Yii::t('CleanThemeModule.config', 'Relative size (em)');
+        $googleFonts =
+            Yii::t('CleanThemeModule.config', 'Google Font name') . ' ' .
+            Button::info(Yii::t('CleanThemeModule.config', 'Google Font name'))->icon('external-link')->link('https://fonts.google.com/')->options(['target' => '_blank'])->loader(false)->xs() . ' (' . Yii::t('CleanThemeModule.config', 'Use the name in the URL') . ')<br>' .
+            Yii::t('CleanThemeModule.config', 'You might need to authorize HumHub to download Google Fonts in the {ContentSecurityPolicy} headers by adding the {googleFontsDownloadUrl} URL after {fontSrcSelf} (see {documentationLink})', [
+                'googleFontsDownloadUrl' => Html::tag('code', 'https://fonts.gstatic.com'),
+                'ContentSecurityPolicy' => Html::tag('code', 'Content-Security-Policy'),
+                'fontSrcSelf' => Html::tag('code', 'font-src \'self\''),
+                'documentationLink' => Button::asLink(Yii::t('CleanThemeModule.config', 'documentation'))->link('https://docs.humhub.org/docs/admin/security#strict-csp-settings')->options(['target' => '_blank']),
+            ]);
+
         return [
-            'containerMaxWidth' => Yii::t('CleanThemeModule.config', 'In pixel'),
-            'fontFamily' => Yii::t('CleanThemeModule.config', 'Google Font name'),
-            'headingFontFamily' => Yii::t('CleanThemeModule.config', 'Google Font name'),
-            'fontSize' => Yii::t('CleanThemeModule.config', 'In pixel'),
-            'phFontSize' => Yii::t('CleanThemeModule.config', 'In pixel'),
-            'rtFontSize' => Yii::t('CleanThemeModule.config', 'In pixel'),
-            'h1FontSize' => Yii::t('CleanThemeModule.config', 'Relative size (em)'),
-            'h1RtFontSize' => Yii::t('CleanThemeModule.config', 'Relative size (em)'),
-            'h2FontSize' => Yii::t('CleanThemeModule.config', 'Relative size (em)'),
-            'h2RtFontSize' => Yii::t('CleanThemeModule.config', 'Relative size (em)'),
-            'h3FontSize' => Yii::t('CleanThemeModule.config', 'Relative size (em)'),
-            'h4FontSize' => Yii::t('CleanThemeModule.config', 'Relative size (em)'),
-            'h5FontSize' => Yii::t('CleanThemeModule.config', 'Relative size (em)'),
-            'h6FontSize' => Yii::t('CleanThemeModule.config', 'Relative size (em)'),
-            'panelBorderWidth' => Yii::t('CleanThemeModule.config', 'In pixel'),
-            'panelBorderRadius' => Yii::t('CleanThemeModule.config', 'In pixel'),
+            'containerMaxWidth' => $inPx,
+            'fontFamily' => $googleFonts,
+            'headingFontFamily' => $googleFonts,
+            'fontSize' => $inPx,
+            'phFontSize' => $inPx,
+            'rtFontSize' => $inPx,
+            'h1FontSize' => $inEm,
+            'h1RtFontSize' => $inEm,
+            'h2FontSize' => $inEm,
+            'h2RtFontSize' => $inEm,
+            'h3FontSize' => $inEm,
+            'h4FontSize' => $inEm,
+            'h5FontSize' => $inEm,
+            'h6FontSize' => $inEm,
+            'panelBorderWidth' => $inPx,
+            'panelBorderRadius' => $inPx,
             'panelBoxShadow' => Yii::t('CleanThemeModule.config', 'CSS value (e.g. {cssValue}, see {documentation})', [
                 'cssValue' => '<code>2px 1px 1px #333;</code>',
                 'documentation' => '<a href="https://developer.mozilla.org/en-US/docs/Web/CSS/box-shadow" target="_blank">' . Yii::t('CleanThemeModule.config', 'documentation') . '</a>',
@@ -232,49 +254,9 @@ class Configuration extends Model
 
     public function loadBySettings(): void
     {
-        $this->containerMaxWidth = (int)$this->settingsManager->get('containerMaxWidth', $this->containerMaxWidth);
-        $this->topMenuNavJustifyContent = $this->settingsManager->get('topMenuNavJustifyContent', $this->topMenuNavJustifyContent);
-        $this->default = $this->settingsManager->get('default', $this->default);
-        $this->primary = $this->settingsManager->get('primary', $this->primary);
-        $this->info = $this->settingsManager->get('info', $this->info);
-        $this->success = $this->settingsManager->get('success', $this->success);
-        $this->warning = $this->settingsManager->get('warning', $this->warning);
-        $this->danger = $this->settingsManager->get('danger', $this->danger);
-        $this->link = $this->settingsManager->get('link', $this->link);
-        $this->menuBorderColor = $this->settingsManager->get('menuBorderColor', $this->menuBorderColor);
-        $this->textColorHeading = $this->settingsManager->get('textColorHeading', $this->textColorHeading);
-        $this->textColorMain = $this->settingsManager->get('textColorMain', $this->textColorMain);
-        $this->textColorDefault = $this->settingsManager->get('textColorDefault', $this->textColorDefault);
-        $this->textColorSecondary = $this->settingsManager->get('textColorSecondary', $this->textColorSecondary);
-        $this->textColorHighlight = $this->settingsManager->get('textColorHighlight', $this->textColorHighlight);
-        $this->textColorSoft = $this->settingsManager->get('textColorSoft', $this->textColorSoft);
-        $this->textColorSoft2 = $this->settingsManager->get('textColorSoft2', $this->textColorSoft2);
-        $this->textColorSoft3 = $this->settingsManager->get('textColorSoft3', $this->textColorSoft3);
-        $this->textColorContrast = $this->settingsManager->get('textColorContrast', $this->textColorContrast);
-        $this->backgroundColorMain = $this->settingsManager->get('backgroundColorMain', $this->backgroundColorMain);
-        $this->backgroundColorSecondary = $this->settingsManager->get('backgroundColorSecondary', $this->backgroundColorSecondary);
-        $this->backgroundColorPage = $this->settingsManager->get('backgroundColorPage', $this->backgroundColorPage);
-        $this->backgroundColorHighlight = $this->settingsManager->get('backgroundColorHighlight', $this->backgroundColorHighlight);
-        $this->fontFamily = $this->settingsManager->get('fontFamily', $this->fontFamily);
-        $this->headingFontFamily = $this->settingsManager->get('headingFontFamily', $this->headingFontFamily);
-        $this->fontSize = (int)$this->settingsManager->get('fontSize', $this->fontSize);
-        $this->phFontSize = (int)$this->settingsManager->get('phFontSize', $this->phFontSize);
-        $this->rtFontSize = (int)$this->settingsManager->get('rtFontSize', $this->rtFontSize);
-        $this->h1FontSize = (float)$this->settingsManager->get('h1FontSize', $this->h1FontSize);
-        $this->h1RtFontSize = (float)$this->settingsManager->get('h1RtFontSize', $this->h1RtFontSize);
-        $this->h2FontSize = (float)$this->settingsManager->get('h2FontSize', $this->h2FontSize);
-        $this->h2RtFontSize = (float)$this->settingsManager->get('h2RtFontSize', $this->h2RtFontSize);
-        $this->h3FontSize = (float)$this->settingsManager->get('h3FontSize', $this->h3FontSize);
-        $this->h4FontSize = (float)$this->settingsManager->get('h4FontSize', $this->h4FontSize);
-        $this->h5FontSize = (float)$this->settingsManager->get('h5FontSize', $this->h5FontSize);
-        $this->h6FontSize = (float)$this->settingsManager->get('h6FontSize', $this->h6FontSize);
-        $this->fontWeight = (int)$this->settingsManager->get('fontWeight', $this->fontWeight);
-        $this->phFontWeight = (int)$this->settingsManager->get('phFontWeight', $this->phFontWeight);
-        $this->panelBorderWidth = (int)$this->settingsManager->get('panelBorderWidth', $this->panelBorderWidth);
-        $this->panelBorderStyle = $this->settingsManager->get('panelBorderStyle', $this->panelBorderStyle);
-        $this->panelBorderColor = $this->settingsManager->get('panelBorderColor', $this->panelBorderColor);
-        $this->panelBorderRadius = (int)$this->settingsManager->get('panelBorderRadius', $this->panelBorderRadius);
-        $this->panelBoxShadow = $this->settingsManager->get('panelBoxShadow', $this->panelBoxShadow);
+        foreach (self::CSS_ATTRIBUTE_UNITS as $name => $unit) {
+            $this->$name = $this->settingsManager->get($name, $this->$name);
+        }
     }
 
     public function save(): bool
@@ -285,123 +267,81 @@ class Configuration extends Model
 
         // If empty value, reset to default value (only for attributes which cannot be empty)
         $defaultConfiguration = new Configuration();
-        $this->containerMaxWidth = $this->containerMaxWidth ?: $defaultConfiguration->containerMaxWidth;
-        $this->topMenuNavJustifyContent = $this->topMenuNavJustifyContent ?: $defaultConfiguration->topMenuNavJustifyContent;
-        $this->default = $this->default ?: $defaultConfiguration->default;
-        $this->primary = $this->primary ?: $defaultConfiguration->primary;
-        $this->info = $this->info ?: $defaultConfiguration->info;
-        $this->success = $this->success ?: $defaultConfiguration->success;
-        $this->warning = $this->warning ?: $defaultConfiguration->warning;
-        $this->danger = $this->danger ?: $defaultConfiguration->danger;
-        $this->link = $this->link ?: $defaultConfiguration->link;
-        $this->menuBorderColor = $this->menuBorderColor ?: $defaultConfiguration->menuBorderColor;
-        $this->textColorHeading = $this->textColorHeading ?: $defaultConfiguration->textColorHeading;
-        $this->textColorMain = $this->textColorMain ?: $defaultConfiguration->textColorMain;
-        $this->textColorDefault = $this->textColorDefault ?: $defaultConfiguration->textColorDefault;
-        $this->textColorSecondary = $this->textColorSecondary ?: $defaultConfiguration->textColorSecondary;
-        $this->textColorHighlight = $this->textColorHighlight ?: $defaultConfiguration->textColorHighlight;
-        $this->textColorSoft = $this->textColorSoft ?: $defaultConfiguration->textColorSoft;
-        $this->textColorSoft2 = $this->textColorSoft2 ?: $defaultConfiguration->textColorSoft2;
-        $this->textColorSoft3 = $this->textColorSoft3 ?: $defaultConfiguration->textColorSoft3;
-        $this->textColorContrast = $this->textColorContrast ?: $defaultConfiguration->textColorContrast;
-        $this->backgroundColorMain = $this->backgroundColorMain ?: $defaultConfiguration->backgroundColorMain;
-        $this->backgroundColorSecondary = $this->backgroundColorSecondary ?: $defaultConfiguration->backgroundColorSecondary;
-        $this->backgroundColorPage = $this->backgroundColorPage ?: $defaultConfiguration->backgroundColorPage;
-        $this->backgroundColorHighlight = $this->backgroundColorHighlight ?: $defaultConfiguration->backgroundColorHighlight;
-        $this->fontSize = $this->fontSize ?: $defaultConfiguration->fontSize;
-        $this->phFontSize = $this->phFontSize ?: $defaultConfiguration->phFontSize;
-        $this->rtFontSize = $this->rtFontSize ?: $defaultConfiguration->rtFontSize;
-        $this->h1FontSize = $this->h1FontSize ?: $defaultConfiguration->h1FontSize;
-        $this->h1RtFontSize = $this->h1RtFontSize ?: $defaultConfiguration->h1RtFontSize;
-        $this->h2FontSize = $this->h2FontSize ?: $defaultConfiguration->h2FontSize;
-        $this->h2RtFontSize = $this->h2RtFontSize ?: $defaultConfiguration->h2RtFontSize;
-        $this->h3FontSize = $this->h3FontSize ?: $defaultConfiguration->h3FontSize;
-        $this->h4FontSize = $this->h4FontSize ?: $defaultConfiguration->h4FontSize;
-        $this->h5FontSize = $this->h5FontSize ?: $defaultConfiguration->h5FontSize;
-        $this->h6FontSize = $this->h6FontSize ?: $defaultConfiguration->h6FontSize;
-        $this->fontWeight = $this->fontWeight ?: $defaultConfiguration->fontWeight;
-        $this->phFontWeight = $this->phFontWeight ?: $defaultConfiguration->phFontWeight;
-        $this->panelBorderWidth = $this->panelBorderWidth ?: $defaultConfiguration->panelBorderWidth;
-        $this->panelBorderStyle = $this->panelBorderStyle ?: $defaultConfiguration->panelBorderStyle;
-        $this->panelBorderColor = $this->panelBorderColor ?: $defaultConfiguration->panelBorderColor;
-        $this->panelBorderRadius = $this->panelBorderRadius ?: $defaultConfiguration->panelBorderRadius;
-
-        // Set values
-        $this->settingsManager->set('containerMaxWidth', $this->containerMaxWidth);
-        $this->settingsManager->set('topMenuNavJustifyContent', $this->topMenuNavJustifyContent);
-        $this->settingsManager->set('default', $this->default);
-        $this->settingsManager->set('primary', $this->primary);
-        $this->settingsManager->set('info', $this->info);
-        $this->settingsManager->set('success', $this->success);
-        $this->settingsManager->set('warning', $this->warning);
-        $this->settingsManager->set('danger', $this->danger);
-        $this->settingsManager->set('link', $this->link);
-        $this->settingsManager->set('menuBorderColor', $this->menuBorderColor);
-        $this->settingsManager->set('textColorHeading', $this->textColorHeading);
-        $this->settingsManager->set('textColorMain', $this->textColorMain);
-        $this->settingsManager->set('textColorDefault', $this->textColorDefault);
-        $this->settingsManager->set('textColorSecondary', $this->textColorSecondary);
-        $this->settingsManager->set('textColorHighlight', $this->textColorHighlight);
-        $this->settingsManager->set('textColorSoft', $this->textColorSoft);
-        $this->settingsManager->set('textColorSoft2', $this->textColorSoft2);
-        $this->settingsManager->set('textColorSoft3', $this->textColorSoft3);
-        $this->settingsManager->set('textColorContrast', $this->textColorContrast);
-        $this->settingsManager->set('backgroundColorMain', $this->backgroundColorMain);
-        $this->settingsManager->set('backgroundColorSecondary', $this->backgroundColorSecondary);
-        $this->settingsManager->set('backgroundColorPage', $this->backgroundColorPage);
-        $this->settingsManager->set('backgroundColorHighlight', $this->backgroundColorHighlight);
-        $this->settingsManager->set('fontFamily', $this->fontFamily);
-        $this->settingsManager->set('headingFontFamily', $this->headingFontFamily);
-        $this->settingsManager->set('fontSize', $this->fontSize);
-        $this->settingsManager->set('phFontSize', $this->phFontSize);
-        $this->settingsManager->set('rtFontSize', $this->rtFontSize);
-        $this->settingsManager->set('h1FontSize', $this->h1FontSize);
-        $this->settingsManager->set('h1RtFontSize', $this->h1RtFontSize);
-        $this->settingsManager->set('h2FontSize', $this->h2FontSize);
-        $this->settingsManager->set('h2RtFontSize', $this->h2RtFontSize);
-        $this->settingsManager->set('h3FontSize', $this->h3FontSize);
-        $this->settingsManager->set('h4FontSize', $this->h4FontSize);
-        $this->settingsManager->set('h5FontSize', $this->h5FontSize);
-        $this->settingsManager->set('h6FontSize', $this->h6FontSize);
-        $this->settingsManager->set('fontWeight', $this->fontWeight);
-        $this->settingsManager->set('phFontWeight', $this->phFontWeight);
-        $this->settingsManager->set('panelBorderWidth', $this->panelBorderWidth);
-        $this->settingsManager->set('panelBorderStyle', $this->panelBorderStyle);
-        $this->settingsManager->set('panelBorderColor', $this->panelBorderColor);
-        $this->settingsManager->set('panelBorderRadius', $this->panelBorderRadius);
-        $this->settingsManager->set('panelBoxShadow', $this->panelBoxShadow);
+        foreach (self::CSS_ATTRIBUTE_UNITS as $name => $unit) {
+            $this->$name = $this->$name ?: $defaultConfiguration->$name;
+            if (static::isFontAttribute($name)) {
+                // Remove + in case the URL of the font was entered
+                $this->$name = str_replace('+', ' ', $this->$name);
+            }
+            $this->settingsManager->set($name, $this->$name);
+        }
 
         $this->generateDynamicCSSFile();
-        Yii::$app->assetManager->clear();
 
         return true;
     }
 
-    private function generateDynamicCSSFile(): void
+    public function generateDynamicCSSFile(): void
     {
-        $css = '/** This file is auto-generated by the Clean Theme configuration. Do not edit the file directly. */' . PHP_EOL . PHP_EOL . ':root {' . PHP_EOL;
+        $css =
+            '/** This file is auto-generated by the Clean Theme configuration. Do not edit the file directly. */' . PHP_EOL . PHP_EOL .
+            ':root {' . PHP_EOL;
 
+        foreach (self::CSS_ATTRIBUTE_UNITS as $name => $unit) {
+            $cssVarName = '--' . Inflector::camel2id($name);
+            $value = static::isFontAttribute($name) ?
+                '"' . $this->$name . '"' :
+                $this->$name;
+
+            // Example for `$containerMaxWidth` attribute: `--container-max-width: 1600px;`
+            $css .= '    ' . $cssVarName . ': ' . $value . $unit . ';' . PHP_EOL;
+        }
+
+        $css .= PHP_EOL;
+
+        // Generate special colors (darkened, lightened and faded colors from the on selected in the configuration)
         foreach ($this->getSpecialColorCssVariables() as $cssVariable) {
             [$amount, $function] = array_reverse(explode('-', $cssVariable));
             $colorName = lcfirst(Inflector::camelize(
                 substr($cssVariable, 2, strlen($cssVariable) - strlen($function . '-' . $amount) - 3)
             ));
-            if ($colorName && $amount && in_array($function, static::SUPPORTED_LESS_FUNCTIONS, true)) {
+            if (
+                $colorName
+                && isset($this->$colorName)
+                && $amount
+                && in_array($function, static::SUPPORTED_LESS_FUNCTIONS, true)
+            ) {
                 $css .= '    ' . $cssVariable . ': ' . ColorHelper::$function($this->$colorName, $amount) . ';' . PHP_EOL;
             }
         }
 
-        $css .= PHP_EOL;
-
-        foreach (self::CSS_ATTRIBUTE_UNITS as $name => $unit) {
-            // Example for `$containerMaxWidth` attribute: `--container-max-width: 1600px;`
-            $css .= '    --' . Inflector::camel2id($name) . ': ' . $this->$name . $unit . ';' . PHP_EOL;
-        }
-
         $css .= '}' . PHP_EOL;
 
+        // Write file
         $dynamicCssFile = Yii::getAlias(self::DYNAMIC_CSS_FILE_PATH . self::DYNAMIC_CSS_FILE_NAME);
         file_put_contents($dynamicCssFile, $css);
+
+        // Clear cache
+        Yii::$app->assetManager->clear();
+    }
+
+    public function getGoogleFontsCss2UrlParams(): string
+    {
+        $fonts = [];
+        if ($this->fontFamily !== 'Open Sans') {
+            $fonts[] = $this->fontFamily;
+        }
+        if (
+            $this->headingFontFamily !== 'Open Sans'
+            && !in_array($this->headingFontFamily, $fonts, true)
+        ) {
+            $fonts[] = $this->headingFontFamily;
+        }
+        $fontsEncoded = [];
+        foreach ($fonts as $font) {
+            $fontsEncoded[] = 'family=' . urlencode($font);
+        }
+        return implode('&', $fontsEncoded);
     }
 
     private function getSpecialColorCssVariables(): array